Using a 'size_t' for the BitCount isn't a good idea - more than 4G of bits overflows (and 4Gb is only ~512MB - pretty small by video standards).  So, use a uint64_t for BitCount.

+++ mjpeg_play/mpeg2enc/elemstrmwriter.hh
@@ -33,7 +33,7 @@ public:
     virtual void WriteOutBufferUpto( const uint8_t *buffer, const uint32_t flush_upto ) = 0;
     inline uint64_t Flushed() const { return flushed; }
     
-    virtual size_t BitCount() = 0;
+    virtual uint64_t BitCount() = 0;
 protected:
     uint64_t flushed;
 };
